pyproject.toml
[project]
name = "telco_cloud_assist_crew"
version = "0.1.0"
description = "telco-cloud-assist-crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "chromadb[http]>=0.5.23",
    "crewai[tools]>=0.105.0,<1.0.0",
    "gradio>=5.21.0",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.19",
    "langchain-ollama>=0.2.3",
    "python-dotenv>=1.0.1",
    "unstructured[docx]>=0.17.0",
]

[project.scripts]
telco_cloud_assist_crew = "telco_cloud_assist_crew.main:run"
run_crew = "telco_cloud_assist_crew.main:run"
train = "telco_cloud_assist_crew.main:train"
replay = "telco_cloud_assist_crew.main:replay"
test = "telco_cloud_assist_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
